refactor: ESLintルールをtestsサブディレクトリに移動
eslint-rules/no-relative-paths-tests.js
→ eslint-rules/tests/no-relative-paths-tests.js
