+++ host-frontend-root/frontend-src-root/eslint-rules/tests/no-relative-paths-tests.js
